login-module/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyJWT
python-dotenv

